--- public/assets/js/server1.js
@@ -1,95 +1,0 @@
-// Dependencies
-// =============================================================
-var express = require("express");
-var path = require("path");
-
-// Sets up the Express App
-// =============================================================
-var app = express();
-var PORT = process.env.PORT || 3000;
-
-// Sets up the Express app to handle data parsing
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-app.use("/public", express.static(__dirname + "/public"));
-
-// Reservations names (DATA)
-// =============================================================
-var tables = [
-  {
-    name: "",
-    phone: "",
-    email: "",
-    uniqueID: ""
-  }
-];
-var waitList = [
-  {
-    name: "",
-    phone: "",
-    email: "",
-    uniqueID: 
-  }
-];
-
-// Routes
-// =============================================================
-
-// Basic route that sends the user first to the AJAX Page
-app.get("/", function(req, res) {
-  res.sendFile(path.join(__dirname, "index.html"));
-});
-
-app.get("/reserve", function(req, res) {
-  res.sendFile(path.join(__dirname, "reserve.html"));
-});
-
-app.get("/tables", function(req, res) {
-    res.sendFile(path.join(__dirname, "tables.html"));
-  });
-
-// Displays all reservations
-app.get("/api/tables", function(req, res) {
-  return res.json(tables);
-});
-
-// Displays a single reservation, or returns false
-app.get("/api/reservations/:reservation", function(req, res) {
-  var chosen = req.params.reservation;
-
-  console.log(chosen);
-
-  for (var i = 0; i < reservations.length; i++) {
-    if (chosen === reservations[i].routeName) {
-      return res.json(reservations[i]);
-    }
-  }
-
-  return res.json(false);
-});
-
-
-
-// Create New names - takes in JSON input
-app.post("/api/reservations", function(req, res) {
-  // req.body hosts is equal to the JSON post sent from the user
-  // This works because of our body parsing middleware
-  var newreservation = req.body;
-
-  // Using a RegEx Pattern to remove spaces from newCharacter
-  // You can read more about RegEx Patterns later https://www.regexbuddy.com/regex.html
-  newreservation.routeName = newreservation.name.replace(/\s+/g, "").toLowerCase();
-
-  console.log(newreservation);
-
-  reservations.push(newreservation);
-
-  res.json(newreservation);
-});
-
-// Starts the server to begin listening
-// =============================================================
-app.listen(PORT, function() {
-  console.log("App listening on PORT " + PORT);
-});
